GF-230: Move RouteColorPicker styles into separate file
